refactor imports, improve UI responsiveness
