+++ frontend/src/App.css
@@ -1,4 +1,5 @@
 #root {
   width: 100vw;
   height: 100vh;
+  overflow: auto;
 }
