+++ frontend/src/components/Modal/ModalContainer.ts
@@ -5,6 +5,8 @@ export const StyledModalContainer = styled(Box)(({ theme }) => ({
   left: '50%',
   transform: 'translate(-50%, -50%)',
   width: "80%",
+  height: "85vh",
+  overflow: "auto",
   background: theme.palette.background.paper,
   border: '2px solid #000',
   boxShadow: "24",
